Factor out buffering of sensor values
